refactor: Mapper update and User Test types
I updated the mapper to use extension methods instead of registering a
mapper to DI. I updated the controllers to use the static methods. I also made the non-admin QuestionController.cs map from the question entity to a user entity suitable for test takers to download. This is so that test takers do not simply inspect the data to get the right answers.