refactor: Introduced UserAnswer type to front end
This introduces the front end to the User Answer type. This manages the
user selection when taking the test.

To make this introduction easier, I also introduced an endpoint that the test can get a list of newed up user answers in the same order that the test questions are in.