refactor: Test form logic extraction
The test form had modal logic making it overly complex. I extracted this
out to its parent and simplified the code greatly. The parameters have
been simplified to a parent callback and to pass a test to edit.

There was also a bug that when you edit a test, it would say to create new. During the headerText parameter extraction, I added extra logic in the parent to set this each time it is going to create/edit.

BREAKING CHANGE: TestForm no longer takes a headerText, isOpen, or
toggle.